move tests for signers into folder called signers
