+++ src/__tests__/signers/NaclSigner.test.ts
@@ -1,4 +1,4 @@
-import NaclSigner from '../signers/NaclSigner'
+import NaclSigner from '../../signers/NaclSigner'
 
 const privateKey = 'nlXR4aofRVuLqtn9+XVQNlX4s1nVQvp+TOhBBtYls1IG+sHyIkDP/WN+rWZHGIQp+v2pyct+rkM4asF/YRFQdQ=='
 const signer = NaclSigner(privateKey)
